requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
WeasyPrint==66.0
